Move MWR response envelope to service boundary
