+++ tests/unit/services/test_mwr_calculation_service.py
@@ -2,8 +2,10 @@ from datetime import date
 
 import pytest
 
+from app.models.mwr_analytics_requests import MoneyWeightedReturnAnalyticsRequest, MWRInputMode
 from app.models.mwr_requests import CashFlow, MoneyWeightedReturnRequest
-from app.services.mwr_calculation_service import calculate_mwr_result
+from app.services.mwr_calculation_service import build_mwr_response, calculate_mwr_result
+from app.services.mwr_mode_service import ResolvedMWRRequest
 from engine.mwr import calculate_money_weighted_return
 
 
@@ -56,3 +58,48 @@ def test_calculate_mwr_result_passes_request_cashflows():
 
     assert request.cash_flows == [CashFlow(amount=10.0, date=date(2025, 6, 30))]
     assert result.method == "MODIFIED_DIETZ"
+
+
+def test_build_mwr_response_preserves_endpoint_payload_contract(mocker):
+    analytics_request = MoneyWeightedReturnAnalyticsRequest.model_validate(
+        {
+            "portfolio_id": "MWR-SERVICE-RESPONSE",
+            "begin_mv": 100.0,
+            "end_mv": 130.0,
+            "as_of": "2025-12-31",
+            "cash_flows": [{"amount": 10.0, "date": "2025-06-30"}],
+            "mwr_method": "DIETZ",
+            "currency": "USD",
+            "report_ccy": "CHF",
+        }
+    )
+    mwr_request = analytics_request.to_stateless_mwr_request()
+    resolved_request = ResolvedMWRRequest(
+        mwr_request=mwr_request,
+        input_mode=MWRInputMode.STATELESS,
+        currency_evidence=None,
+    )
+    mwr_result = calculate_mwr_result(mwr_request)
+    supportability_metric = mocker.patch("app.services.mwr_calculation_service.record_supportability_metric")
+    solver_metric = mocker.patch("app.services.mwr_calculation_service.record_mwr_solver_outcome")
+
+    response = build_mwr_response(
+        request=analytics_request,
+        resolved_request=resolved_request,
+        mwr_result=mwr_result,
+        input_fingerprint="input-fingerprint",
+        calculation_hash="calculation-hash",
+        engine_version="test-version",
+    )
+
+    assert response.portfolio_id == "MWR-SERVICE-RESPONSE"
+    assert response.reporting_currency == "CHF"
+    assert response.cashflows_used == [CashFlow(amount=10.0, date=date(2025, 6, 30))]
+    assert response.meta.engine_version == "test-version"
+    assert response.meta.input_fingerprint == "input-fingerprint"
+    assert response.meta.calculation_hash == "calculation-hash"
+    assert response.diagnostics.effective_period_start == mwr_result.start_date
+    assert response.audit.counts == {"cashflows": 1}
+    assert response.calculation_supportability.resolved_period_count == 1
+    supportability_metric.assert_called_once()
+    solver_metric.assert_called_once()
